LayoutTests/platform/chromium-mac-snowleopard/accessibility/aria-checkbox-sends-notification-expected.txt

Test Checkbox
This tests that checking of an aria checkbox sends a notification.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Got notification: CheckedStateChanged
Got notification: CheckedStateChanged
Got notification: LayoutComplete
Got notification: LayoutComplete
PASS successfullyParsed is true

TEST COMPLETE
